+++ Models/BrandAmount.cs
@@ -9,11 +9,11 @@ namespace VaccineAPI.Models
     {
         public long Id { get; set; }
         public int Amount { get; set; }
-        public int BrandId { get; set; }
+        public long BrandId { get; set; }
          [JsonIgnore]
         public Brand Brand { get; set; }
     
-        public int DoctorId { get; set; }
+        public long DoctorId { get; set; }
          [JsonIgnore]
         public Doctor Doctor { get; set; }
     }
